+++ runtime/browser/android/xwalk_http_auth_handler.cc
@@ -28,7 +28,7 @@ XWalkHttpAuthHandler::XWalkHttpAuthHandler(XWalkLoginDelegate* login_delegate,
   JNIEnv* env = base::android::AttachCurrentThread();
   http_auth_handler_.Reset(
       Java_XWalkHttpAuthHandler_create(
-          env, reinterpret_cast<jint>(this), first_auth_attempt));
+          env, reinterpret_cast<intptr_t>(this), first_auth_attempt));
 }
 
 XWalkHttpAuthHandler:: ~XWalkHttpAuthHandler() {
